Refactor and rename media grouping component
